requirements.txt
streamlit
datasets
sentence-transformers
transformers
faiss-cpu
torch
